requirements.txt
autopep8==2.0.2
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cryptography==41.0.2
fernet==1.0.1
Flask==2.3.2
Flask-Cors==4.0.0
gcloud==0.18.3
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
libmagic==1.0
MarkupSafe==2.1.3
oauth2client==4.1.3
protobuf==4.24.0rc1
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pyparsing==3.1.0
Pyrebase4==4.7.1
python-dotenv==1.0.0
python-http-client==3.3.7
python-jwt==2.0.1
python-magic==0.4.27
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
sendgrid==6.10.0
six==1.16.0
starkbank-ecdsa==2.2.0
urllib3==1.26.16
Werkzeug==2.3.6
